+++ apps/pie-storybook/stories/pie-switch.stories.ts
@@ -177,5 +177,19 @@ const createSwitchStory = createStory(Template, defaultArgs);
 
 const createSwitchStoryWithForm = createStory<SwitchProps>(FormTemplate, defaultArgs);
 
-export const Default = createSwitchStory();
+const formIntegrationOnly = {
+    description: 'This prop is only used when the switch is used inside a form. You can try it out in the Form Integration story.',
+    table: {
+        readonly: true,
+    },
+};
+
+export const Default = createSwitchStory({}, {
+    argTypes: {
+        name: formIntegrationOnly,
+        required: formIntegrationOnly,
+        value: formIntegrationOnly,
+    },
+});
+
 export const FormIntegration = createSwitchStoryWithForm({ label: 'Click me' });