+++ packages/components/pie-cookie-banner/test/accessibility/pie-cookie-banner.spec.ts
@@ -1,15 +1,11 @@
 
-import { test, expect } from '@justeattakeaway/pie-webc-testing/src/playwright/webc-fixtures.ts';
-import { PieCookieBanner, type CookieBannerProps } from '../../src/index.ts';
+import { test, expect } from '@justeattakeaway/pie-webc-testing/src/playwright/playwright-fixtures.ts';
+import { CookieBannerComponent } from '../helpers/page-object/pie-cookie-banner.page.ts';
 
 test.describe('PieCookieBanner - Accessibility tests', () => {
-    test('a11y - should test the PieCookieBanner component WCAG compliance', async ({ makeAxeBuilder, mount }) => {
-        await mount(
-            PieCookieBanner,
-            {
-                props: {} as CookieBannerProps,
-            },
-        );
+    test('a11y - should test the PieCookieBanner component WCAG compliance', async ({ makeAxeBuilder, page }) => {
+        const cookieBannerPage = new CookieBannerComponent(page);
+        await cookieBannerPage.load();
 
         const results = await makeAxeBuilder().analyze();