+++ packages/components/pie-divider/src/index.ts
@@ -1,9 +1,10 @@
 import { LitElement, html, unsafeCSS } from 'lit';
 import { property } from 'lit/decorators.js';
-import { validPropertyValues, defineCustomElement } from '@justeattakeaway/pie-webc-core';
+import { classMap } from 'lit/directives/class-map.js';
+import { defineCustomElement, validPropertyValues } from '@justeattakeaway/pie-webc-core';
 import styles from './divider.scss?inline';
 import {
-    type DividerProps, variants, orientations, defaultProps,
+    type DividerProps, defaultProps, orientations, variants,
 } from './defs';
 
 // Valid values available to consumers
@@ -17,23 +18,26 @@ const componentSelector = 'pie-divider';
 export class PieDivider extends LitElement implements DividerProps {
     @property({ type: String })
     @validPropertyValues(componentSelector, variants, defaultProps.variant)
-    public variant: DividerProps['variant'] = defaultProps.variant;
+    public variant = defaultProps.variant;
 
     @property({ type: String })
     @validPropertyValues(componentSelector, orientations, defaultProps.orientation)
-    public orientation : DividerProps['orientation'] = defaultProps.orientation;
+    public orientation = defaultProps.orientation;
 
     render () {
         const { variant, orientation } = this;
+
+        const classes = {
+            'c-divider': true,
+            'c-divider--inverse': variant === 'inverse',
+            'c-divider--vertical': orientation === 'vertical',
+        };
 
         return html`
             <hr
                 data-test-id="pie-divider"
                 aria-hidden="true"
-                class="c-divider"
-                variant=${variant}
-                orientation=${orientation}
-            />`;
+                class="${classMap(classes)}" />`;
     }
 
     // Renders a `CSSResult` generated from SCSS by Vite