+++ packages/components/pie-textarea/src/index.ts
@@ -1,38 +1,42 @@
 import {
     LitElement, html, unsafeCSS, type PropertyValues, nothing,
 } from 'lit';
-
 import { property, query } from 'lit/decorators.js';
+import { classMap } from 'lit/directives/class-map.js';
+import { ifDefined } from 'lit/directives/if-defined.js';
 import { live } from 'lit/directives/live.js';
 import throttle from 'lodash.throttle';
 
+import '@justeattakeaway/pie-assistive-text';
+import '@justeattakeaway/pie-form-label';
 import {
-    validPropertyValues, RtlMixin, defineCustomElement, FormControlMixin, wrapNativeEvent,
+    validPropertyValues, RtlMixin, defineCustomElement, FormControlMixin, wrapNativeEvent, type PIEInputElement,
 } from '@justeattakeaway/pie-webc-core';
 
-import { ifDefined } from 'lit/directives/if-defined.js';
 import styles from './textarea.scss?inline';
 import {
-    type TextareaProps, defaultProps, sizes, resizeModes,
+    type TextareaProps, defaultProps, sizes, resizeModes, statusTypes,
 } from './defs';
-
-import '@justeattakeaway/pie-form-label';
 
 // Valid values available to consumers
 export * from './defs';
 
 const componentSelector = 'pie-textarea';
+const assistiveTextIdValue = 'assistive-text';
 
 /**
  * @tagname pie-textarea
  * @event {InputEvent} input - when the textarea value is changed.
  * @event {CustomEvent} change - when the textarea value is changed.
  */
-export class PieTextarea extends FormControlMixin(RtlMixin(LitElement)) implements TextareaProps {
+export class PieTextarea extends FormControlMixin(RtlMixin(LitElement)) implements TextareaProps, PIEInputElement {
     static shadowRootOptions = { ...LitElement.shadowRootOptions, delegatesFocus: true };
 
     @property({ type: String })
     public value = defaultProps.value;
+
+    @property({ type: String })
+    public defaultValue: TextareaProps['defaultValue'];
 
     @property({ type: Boolean, reflect: true })
     public disabled = defaultProps.disabled;
@@ -61,18 +65,28 @@ export class PieTextarea extends FormControlMixin(RtlMixin(LitElement)) implemen
     public required = defaultProps.required;
 
     @property({ type: String })
-    public name?: TextareaProps['name'];
+    @validPropertyValues(componentSelector, statusTypes, defaultProps.status)
+    public status = defaultProps.status;
 
     @property({ type: String })
-    public autocomplete?: TextareaProps['autocomplete'];
+    public assistiveText: TextareaProps['assistiveText'];
+
+    @property({ type: String })
+    public name: TextareaProps['name'];
+
+    @property({ type: String })
+    public autocomplete: TextareaProps['autocomplete'];
 
     @query('textarea')
     private _textarea!: HTMLTextAreaElement;
 
+    @query('textarea')
+    public focusTarget!: HTMLElement;
+
     private _throttledResize = throttle(() => {
         if (this.resize === 'auto') {
             this._textarea.style.height = 'auto';
-            this._textarea.style.height = `${this._textarea.scrollHeight + 2}px`; // +2 for border thicknesses
+            this._textarea.style.height = `${this._textarea.scrollHeight}px`;
         }
     }, 100);
 
@@ -99,7 +113,7 @@ export class PieTextarea extends FormControlMixin(RtlMixin(LitElement)) implemen
      * Resets the value to the default value.
      */
     public formResetCallback (): void {
-        this.value = defaultProps.value;
+        this.value = this.defaultValue ?? defaultProps.value;
 
         this._internals.setFormValue(this.value);
     }
@@ -108,6 +122,7 @@ export class PieTextarea extends FormControlMixin(RtlMixin(LitElement)) implemen
         this.restrictInputLength();
         this._internals.setFormValue(this.value);
 
+        window.addEventListener('resize', () => this.handleResize());
         this._textarea.addEventListener('keydown', this.handleKeyDown);
     }
 
@@ -166,6 +181,7 @@ export class PieTextarea extends FormControlMixin(RtlMixin(LitElement)) implemen
 
     public disconnectedCallback (): void {
         this._textarea.removeEventListener('keydown', this.handleKeyDown);
+        window.removeEventListener('resize', () => this.handleResize());
     }
 
     renderLabel (label: string, maxLength?: number) {
@@ -174,6 +190,21 @@ export class PieTextarea extends FormControlMixin(RtlMixin(LitElement)) implemen
         return label?.length
             ? html`<pie-form-label for="${componentSelector}" trailing=${ifDefined(characterCount)}>${label}</pie-form-label>`
             : nothing;
+    }
+
+    renderAssistiveText () {
+        if (!this.assistiveText) {
+            return nothing;
+        }
+
+        return html`
+            <pie-assistive-text
+                id="${assistiveTextIdValue}"
+                variant=${ifDefined(this.status)}
+                data-test-id="pie-textarea-assistive-text">
+                ${this.assistiveText}
+            </pie-assistive-text>
+        `;
     }
 
     render () {
@@ -189,15 +220,23 @@ export class PieTextarea extends FormControlMixin(RtlMixin(LitElement)) implemen
             required,
             label,
             maxLength,
+            status,
+            assistiveText,
         } = this;
 
-        return html`
+        const classes = {
+            'c-textareaWrapper': true,
+            'c-textarea--disabled': disabled,
+            'c-textarea--error': status === 'error',
+            [`c-textarea--resize-${resize}`]: true,
+            [`c-textarea--${size}`]: true,
+        };
+
+        return html`<div>
+            ${this.renderLabel(label, maxLength)}
             <div
-                class="c-textareaWrapper"
-                data-test-id="pie-textarea-wrapper"
-                data-pie-size="${size}"
-                data-pie-resize="${resize}">
-                ${this.renderLabel(label, maxLength)}
+                class="${classMap(classes)}"
+                data-test-id="pie-textarea-wrapper">
                 <textarea
                     id="${componentSelector}"
                     data-test-id="${componentSelector}"
@@ -208,10 +247,15 @@ export class PieTextarea extends FormControlMixin(RtlMixin(LitElement)) implemen
                     ?readonly=${readonly}
                     ?required=${required}
                     ?disabled=${disabled}
+                    aria-describedby=${ifDefined(assistiveText ? assistiveTextIdValue : undefined)}
+                    aria-invalid=${status === 'error' ? 'true' : 'false'}
+                    aria-errormessage="${ifDefined(status === 'error' ? assistiveTextIdValue : undefined)}"
                     @input=${this.handleInput}
                     @change=${this.handleChange}
                 ></textarea>
-            </div>`;
+            </div>
+            ${this.renderAssistiveText()}
+        </div>`;
     }
 
     // Renders a `CSSResult` generated from SCSS by Vite