+++ packages/components/pie-assistive-text/CHANGELOG.md
@@ -1,5 +1,22 @@
 # @justeattakeaway/pie-assistive-text
 
+## 0.7.0
+
+### Minor Changes
+
+- [Changed] - Use classmap for assistive-text variant styles ([#1694](https://github.com/justeattakeaway/pie/pull/1694)) by [@jamieomaguire](https://github.com/jamieomaguire)
+
+## 0.6.1
+
+### Patch Changes
+
+- [Added] - pie-css as a dev dependency ([#1667](https://github.com/justeattakeaway/pie/pull/1667)) by [@xander-marjoram](https://github.com/xander-marjoram)
+
+  [Changed] - updated component generator
+
+- Updated dependencies [[`ccfe4583c`](https://github.com/justeattakeaway/pie/commit/ccfe4583c274531d879b9c877f4c5c08dc9418cd)]:
+  - @justeattakeaway/pie-icons-webc@0.25.0
+
 ## 0.6.0
 
 ### Minor Changes